Join tests into one file
